Backend/requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
pillow
python-dotenv
requests
chromadb
PyPDF2
python-docx